Improved: Data Migration by releases wiki page
Pandoc was accurately not rendering wrong MarkDown code.

Anyway this needs to be translater to Asciidoc one day or another. 

There is a Confluence plugin to import Asciidoc, not sure we will want to do 
that or only keep the doc in the source. I guess it depends on the importance
of the document. Some needs really to be visible in wiki, or maybe not and why
keeping the wiki at all ?


git-svn-id: http://svn.apache.org/repos/asf/ofbiz/tools@1829134 13f79535-47bb-0310-9956-ffa450edef68