Implemented: Remove the Gradle wrapper from our release packages and add a step to our build notes (OFBIZ-10145)
As explained in the Jira, copy the OFBiz/gradle/wrapper directory in 
tools/buildbot/gradle/wrapper.

Later in Buildbot copy it back to its place, et voilà :)

When we will update Gradle we will need to update those. Anyway we need to ask
infra for the same. We have Gradle installed in /usr/lib/gradle/ on 
"lares_ubuntu" , "orcus_ubuntu" , "silvanus_ubuntu"  as says INFRA-13202.
Actually It's about Gradle to 3.2.1, I guess it's the same for Gradle 5.0.

I still need to modify the Buildbot conf file at 
https://svn.apache.org/repos/infra/infrastructure/buildbot/aegis/buildmaster/master1/projects/ofbiz.conf

git-svn-id: http://svn.apache.org/repos/asf/ofbiz/tools@1860627 13f79535-47bb-0310-9956-ffa450edef68