Improved: no functional change
As discussed with Mathieu on dev ML change the location of  meta-data files that 
are maintained manually. They are distributed inside the JAR to allow the 
‘ServiceLoader’ API [1] to find the classes implementing a particular interface 
(corresponding to the file name) efficiently.
[1]https://docs.oracle.com/javase/8/docs/api/java/util/ServiceLoader.html

I spotted this with RAT, for base.util.test.UtilObjectTests$TestFactoryIntf 
The other same files have ASL2 headers, I decided to not change that.

git-svn-id: http://svn.apache.org/repos/asf/ofbiz/tools@1866514 13f79535-47bb-0310-9956-ffa450edef68