Improved: no functional change
Adds new Rat detected issues which are not

.sonarcloud.properties is used to run sonar, see OFBIZ-10917

As discussed previously with Mathieu on dev ML change the location of  meta-data
files that are maintained manually. They are distributed inside the JAR to allow
the  ‘ServiceLoader’ API [1] to find the classes implementing a particular
interface (corresponding to the file name) efficiently.
[1]https://docs.oracle.com/javase/8/docs/api/java/util/ServiceLoader.html

I spotted this with RAT, for base.util.test.UtilObjectTests$TestFactoryIntf
The other same files have ASL2 headers, I decided to not change that.

I wonder if my previous commit about
**/framework/base/src/main/resources/META-INF/**
was right. I was unable to test then because Buildbot was unavailable.
I replace it by
**/framework/base/src/test/resources/META-INF/**