Improved: handle patches in demos
Nicolas started a good work 6 months ago and I think it should be continued.
Since then we have switched to Git so his files are still relevant but need to
be adapted.

In the meantime I'll handle the patches in a simple way. I have moved the
patches from the main dir to a patch dir, as did Nicolas, and removed the patch
dir under Nicolas's dir

This also adds the solr.config.patch provided by Ravi for OFBIZ-11504