Improved: handle patches in demos
I have decided to apply patches once and for all. The reason is else they are
applied once and not later. So it's easier like that.

If we need to change the patches they will be reverted and applied again.
Hopefully only when changing stable and old.

But in R16 (old demo, still uses specialpurpose) the ones under plugins must be
applied  each time because pullAllPluginsSource removes the plugins dir, hence
the patched files