Improved: Clean OFBizJavaFormatter.xml (OFBIZ-11807)
When I initially committed OFBizJavaFormatter.xml at
http://svn.apache.org/viewvc?view=revision&revision=1819045 I did not notice
that I committed not only 1 but 3 slightly different formatters. The ones that I
had then on my machine. Those have persisted since today and it does not make
sense to have 3, it's confusing.

Unfortunately it's hard to decide which one to pick because the documentation is
sometimes abstruse: https://s.apache.org/ad0hr

After comparing them and knowing that checkstyle.xml does not reflect the same
I finally decided to make things simple and to keep only the 1st one named as
the file: OFBizJavaFormatter.xml