Updated README to reference docker for demo sites (OFBIZ-12863)
Added references to the 3 experimental hostnames that can be hosted by the demo VM.

Removed references to an upgrade process since docker image building is handled in GitHub workflows.

Added brief description of the daily cron job used to pull updated docker images used for the demo sites.

Added reference to the ofbizdocker README file which goes into greater detail on how Docker is used to host the demo sites.